requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
drf-nested-routers==0.92.1
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.4.1
djangogrpcframework==0.2
grpcio==1.33.2
grpcio-tools==1.33.2
